refactor(scene3d): 抽出场景视图/编排/取景预览,壳出白名单 [split 5/5]
- scene3dSceneView.tsx: CameraFrustumLines/CameraTargetFeedback/SceneObjectView/CameraHelperView
- scene3dSceneContent.tsx: SceneContent(场景编排器)
- scene3dCameraPreview.tsx: cameraPreviewViewportStyle/CameraPreviewPose/PreviewObjectView/CameraPreviewScene/CameraPreview
- 清壳 ~50 个死 import;Scene3DFullscreen 3822→771 < 800,移出 check-file-sizes 白名单(巨壳债清一个)

3822 行最大巨壳 → 771 壳 + 7 子模块全部 < 800。双 tsc 零错误,画布 551 测试绿
