+++ src/workbench/generationCanvas/nodes/scene3d/Scene3DFullscreen.tsx
@@ -1,19 +1,11 @@
 import React from 'react'
 import { createPortal } from 'react-dom'
-import { Canvas, type ThreeEvent, useFrame, useThree } from '@react-three/fiber'
+import { Canvas } from '@react-three/fiber'
 import { AnimatePresence, motion } from 'framer-motion'
-import {
-  Environment,
-  Grid,
-  OrbitControls,
-  Sky,
-  TransformControls,
-} from '@react-three/drei'
 import {
   IconArrowsMove,
   IconCamera,
   IconCube,
-  IconEye,
   IconListTree,
   IconPhoto,
   IconRotate,
@@ -21,16 +13,10 @@ import {
   IconWorld,
   IconX,
 } from '@tabler/icons-react'
-import * as THREE from 'three'
-import { cn } from '../../../../utils/cn'
 import { toast } from '../../../../ui/toast'
 import { cloneScene3DState } from './scene3dSerializer'
-import { CameraStateRecorder } from './CameraStateRecorder'
 import {
-  SCENE3D_ASPECT_OPTIONS,
-  SCENE3D_ASPECT_RATIOS,
   type CaptureApi,
-  type Scene3DAspectRatio,
   type Scene3DCamera,
   type Scene3DCaptureResult,
   type Scene3DControlMode,
@@ -41,47 +27,20 @@ import {
   type Scene3DTransformMode,
   type Scene3DVector3,
 } from './scene3dTypes'
-
 import {
   OBJECT_LIMIT,
-  CAMERA_HELPER_FLAG,
-  SCENE3D_GRID_FLAG,
   FULLSCREEN_Z_INDEX,
-  CAMERA_MARKER_COLOR,
-  CAMERA_MARKER_ACCENT_COLOR,
-  CAMERA_HELPER_VISUAL_FAR,
-  CAMERA_AIM_FEEDBACK_LENGTH,
-  CAMERA_AIM_HANDLE_DISTANCE,
-  CAMERA_DEFAULT_TARGET,
-  FREE_LOOK_ROTATION_SPEED,
-  WHEEL_TRAVEL_SPEED,
-  GRID_CELL_COLOR,
-  GRID_SECTION_COLOR,
-  DARK_GRID_CELL_COLOR,
-  DARK_GRID_SECTION_COLOR,
   type CrowdAddOptions,
-  type Scene3DMovementCode,
 } from './scene3dConstants'
 import { PanelButton, CanvasPanelRestoreButton, SceneAddToolbar } from './scene3dToolbar'
 import {
   isEditableKeyboardTarget,
-  pointerCaptureTarget,
-  vectorFromArray,
-  vectorToArray,
   cameraLookAtRotation,
   levelEditorCameraRotation,
   applyEditorCameraPose,
-  applySceneCameraPose,
   editorCameraFromSceneCamera,
-  eulerToArray,
   vectorAlmostEqual,
-  cameraPoseSampleChanged,
-  cameraAimSpherical,
-  type CameraPoseSample,
   crowdCount,
-  aspectDimensions,
-  captureScene,
-  mannequinRoleLabel,
   makeObject,
   makeCrowdObject,
   makeCamera,
@@ -89,36 +48,11 @@ import {
   cloneCameraForClipboard,
   makePastedObject,
   makePastedCamera,
-  isMovementCode,
-  clearMovementKeyState,
-  hasActiveMovementKey,
-  type PointerCaptureTarget,
 } from './scene3dMath'
 import { SceneObjectList, PropertyPanel } from './scene3dInspector'
-import {
-  Scene3DMeshGeometry,
-  ProceduralMannequin,
-  MannequinAssetBoundary,
-  Mannequin,
-  MannequinCrowd,
-  ProceduralMannequinCrowd,
-  LightObject,
-  MannequinRoleLabel,
-  singleMannequinLabelPosition,
-  crowdLabelPositions,
-  objectGroundFootprint,
-  objectVisualHalfHeight,
-  objectTransformAnchorPosition,
-  nextAvailableObjectPosition,
-  MannequinFootRings,
-} from './scene3dObjects'
-import {
-  Scene3DControls,
-  InitialCameraPose,
-  FocusController,
-  CaptureBinder,
-  CameraViewEditController,
-} from './scene3dViewControllers'
+import { nextAvailableObjectPosition } from './scene3dObjects'
+import { SceneContent } from './scene3dSceneContent'
+import { CameraPreview } from './scene3dCameraPreview'
 
 type Scene3DFullscreenProps = {
   initialState: Scene3DState
@@ -133,974 +67,6 @@ type Scene3DFullscreenProps = {
 type Scene3DClipboardItem =
   | { type: 'object'; item: Scene3DObject; pasteCount: number }
   | { type: 'camera'; item: Scene3DCamera; pasteCount: number }
-function CameraFrustumLines({
-  cameraData,
-  selected,
-}: {
-  cameraData: Scene3DCamera
-  selected: boolean
-}): JSX.Element {
-  const positions = React.useMemo(() => {
-    const distance = Math.min(cameraData.far, Math.max(cameraData.near + 0.1, CAMERA_HELPER_VISUAL_FAR))
-    const aspect = SCENE3D_ASPECT_RATIOS[cameraData.aspectRatio]
-    const halfHeight = Math.tan(THREE.MathUtils.degToRad(cameraData.fov) / 2) * distance
-    const halfWidth = halfHeight * aspect
-    const origin: Scene3DVector3 = [0, 0, 0]
-    const topLeft: Scene3DVector3 = [-halfWidth, halfHeight, distance]
-    const topRight: Scene3DVector3 = [halfWidth, halfHeight, distance]
-    const bottomRight: Scene3DVector3 = [halfWidth, -halfHeight, distance]
-    const bottomLeft: Scene3DVector3 = [-halfWidth, -halfHeight, distance]
-    const segments = [
-      origin, topLeft,
-      origin, topRight,
-      origin, bottomRight,
-      origin, bottomLeft,
-      topLeft, topRight,
-      topRight, bottomRight,
-      bottomRight, bottomLeft,
-      bottomLeft, topLeft,
-    ]
-    return new Float32Array(segments.flat())
-  }, [cameraData.aspectRatio, cameraData.far, cameraData.fov, cameraData.near])
-
-  return (
-    <lineSegments frustumCulled={false} raycast={() => null} userData={{ [CAMERA_HELPER_FLAG]: true }}>
-      <bufferGeometry>
-        <bufferAttribute attach="attributes-position" args={[positions, 3]} />
-      </bufferGeometry>
-      <lineBasicMaterial
-        color={selected ? '#facc15' : '#64748b'}
-        opacity={selected ? 0.9 : 0.56}
-        transparent
-        toneMapped={false}
-      />
-    </lineSegments>
-  )
-}
-
-function CameraTargetFeedback({ cameraData }: { cameraData: Scene3DCamera }): JSX.Element {
-  const target = cameraData.target || CAMERA_DEFAULT_TARGET
-  const endpoint = React.useMemo(() => {
-    const position = vectorFromArray(cameraData.position)
-    const direction = vectorFromArray(target).sub(position)
-    if (direction.lengthSq() < 0.0001) direction.set(0, 0, 1)
-    direction.normalize().multiplyScalar(CAMERA_AIM_FEEDBACK_LENGTH)
-    return vectorToArray(position.add(direction))
-  }, [cameraData.position, target])
-  const positions = React.useMemo(() => new Float32Array([
-    cameraData.position[0],
-    cameraData.position[1],
-    cameraData.position[2],
-    endpoint[0],
-    endpoint[1],
-    endpoint[2],
-  ]), [cameraData.position, endpoint])
-
-  return (
-    <>
-      <lineSegments frustumCulled={false} raycast={() => null} userData={{ [CAMERA_HELPER_FLAG]: true }}>
-        <bufferGeometry>
-          <bufferAttribute attach="attributes-position" args={[positions, 3]} />
-        </bufferGeometry>
-        <lineBasicMaterial color="#facc15" opacity={0.62} transparent toneMapped={false} />
-      </lineSegments>
-      <mesh position={endpoint} raycast={() => null} userData={{ [CAMERA_HELPER_FLAG]: true }}>
-        <sphereGeometry args={[0.055, 18, 12]} />
-        <meshBasicMaterial color="#facc15" toneMapped={false} />
-      </mesh>
-    </>
-  )
-}
-
-function SceneObjectView({
-  object,
-  selected,
-  readOnly,
-  transformMode,
-  orbitControlsActive,
-  navigationLockedRef,
-  roleLabel,
-  roleStartIndex,
-  onSelect,
-  onFocus,
-  onTransformStart,
-  onTransformEnd,
-  onTransform,
-}: {
-  object: Scene3DObject
-  selected: boolean
-  readOnly: boolean
-  transformMode: Scene3DTransformMode
-  orbitControlsActive: boolean
-  navigationLockedRef: React.MutableRefObject<boolean>
-  roleLabel?: string
-  roleStartIndex?: number
-  onSelect: () => void
-  onFocus: () => void
-  onTransformStart: () => void
-  onTransformEnd: () => void
-  onTransform: (patch: Partial<Scene3DObject>) => void
-}): JSX.Element {
-  const visualRef = React.useRef<THREE.Group>(null!) as React.MutableRefObject<THREE.Group>
-  const anchorRef = React.useRef<THREE.Group>(null!) as React.MutableRefObject<THREE.Group>
-  const transformRef = React.useRef<any>(null)
-  const transformDraggingRef = React.useRef(false)
-  const orbitControlsActiveRef = React.useRef(orbitControlsActive)
-  const { controls } = useThree()
-  const anchorPosition = React.useMemo(() => objectTransformAnchorPosition(object), [object])
-
-  const handleObjectChange = React.useCallback(() => {
-    if (!anchorRef.current) return
-    const nextScale = vectorToArray(anchorRef.current.scale)
-    const nextPosition: Scene3DVector3 = [
-      Number(anchorRef.current.position.x.toFixed(4)),
-      Number((anchorRef.current.position.y + objectVisualHalfHeight(object, nextScale)).toFixed(4)),
-      Number(anchorRef.current.position.z.toFixed(4)),
-    ]
-    const nextRotation = eulerToArray(anchorRef.current.rotation)
-    if (visualRef.current) {
-      visualRef.current.position.fromArray(nextPosition)
-      visualRef.current.rotation.copy(anchorRef.current.rotation)
-      visualRef.current.scale.copy(anchorRef.current.scale)
-    }
-    onTransform({
-      position: nextPosition,
-      rotation: nextRotation,
-      scale: nextScale,
-    })
-  }, [object, onTransform])
-
-  React.useLayoutEffect(() => {
-    orbitControlsActiveRef.current = orbitControlsActive
-    if (!orbitControlsActive && controls && 'enabled' in controls && !transformDraggingRef.current) {
-      ;(controls as { enabled: boolean }).enabled = false
-    }
-  }, [controls, orbitControlsActive])
-
-  React.useLayoutEffect(() => {
-    if (!anchorRef.current || transformDraggingRef.current) return
-    anchorRef.current.position.fromArray(anchorPosition)
-    anchorRef.current.rotation.fromArray(object.rotation)
-    anchorRef.current.scale.fromArray(object.scale)
-  }, [anchorPosition, object.rotation, object.scale])
-
-  React.useEffect(() => {
-    const tc = transformRef.current
-    if (!tc) return
-    const handler = (event: any) => {
-      const dragging = Boolean(event.value)
-      const wasDragging = transformDraggingRef.current
-      transformDraggingRef.current = dragging
-      navigationLockedRef.current = dragging
-      if (dragging && !wasDragging) {
-        orbitControlsActiveRef.current = false
-        onTransformStart()
-      }
-      if (controls && 'enabled' in controls) {
-        ;(controls as { enabled: boolean }).enabled = dragging ? false : orbitControlsActiveRef.current
-      }
-    }
-    tc.addEventListener('dragging-changed', handler)
-    return () => {
-      if (transformDraggingRef.current) {
-        navigationLockedRef.current = false
-        transformDraggingRef.current = false
-        onTransformEnd()
-      }
-      tc.removeEventListener('dragging-changed', handler)
-    }
-  }, [controls, navigationLockedRef, onTransformEnd, onTransformStart, selected])
-
-  const handleTransformMouseDown = React.useCallback(() => {
-    orbitControlsActiveRef.current = false
-    navigationLockedRef.current = true
-    onTransformStart()
-    if (controls && 'enabled' in controls) {
-      ;(controls as { enabled: boolean }).enabled = false
-    }
-  }, [controls, navigationLockedRef, onTransformStart])
-
-  const handleTransformMouseUp = React.useCallback(() => {
-    navigationLockedRef.current = false
-    onTransformEnd()
-    if (controls && 'enabled' in controls) {
-      ;(controls as { enabled: boolean }).enabled = orbitControlsActiveRef.current
-    }
-  }, [controls, navigationLockedRef, onTransformEnd])
-
-  const group = (
-    <group
-      ref={visualRef}
-      visible={object.visible}
-      position={object.position}
-      rotation={object.rotation}
-      scale={object.scale}
-      onPointerDown={(event) => {
-        event.stopPropagation()
-        onSelect()
-      }}
-      onDoubleClick={(event) => {
-        event.stopPropagation()
-        onSelect()
-        onFocus()
-      }}
-    >
-      {object.type === 'mannequin' ? (
-        <MannequinAssetBoundary fallback={<ProceduralMannequin color={object.color || '#808080'} />}>
-          <React.Suspense fallback={<ProceduralMannequin color={object.color || '#808080'} />}>
-            <Mannequin color={object.color || '#808080'} pose={object.pose} />
-          </React.Suspense>
-        </MannequinAssetBoundary>
-      ) : object.type === 'mannequinCrowd' ? (
-        <MannequinAssetBoundary fallback={<ProceduralMannequinCrowd object={object} roleStartIndex={roleStartIndex || 0} />}>
-          <React.Suspense fallback={<ProceduralMannequinCrowd object={object} roleStartIndex={roleStartIndex || 0} />}>
-            <MannequinCrowd object={object} roleStartIndex={roleStartIndex || 0} />
-          </React.Suspense>
-        </MannequinAssetBoundary>
-      ) : object.type === 'light' ? (
-        <>
-          <LightObject object={object} />
-          <mesh>
-            <sphereGeometry args={[0.12, 18, 12]} />
-            <meshBasicMaterial color={object.lightColor || '#ffffff'} toneMapped={false} />
-          </mesh>
-        </>
-      ) : (
-        <mesh>
-          <Scene3DMeshGeometry geometry={object.geometry} />
-          <meshStandardMaterial
-            color={object.color || '#808080'}
-            roughness={0.55}
-            metalness={0.04}
-            side={object.geometry === 'plane' ? THREE.DoubleSide : THREE.FrontSide}
-          />
-        </mesh>
-      )}
-      {object.type === 'mannequinCrowd' ? (
-        <mesh>
-          <boxGeometry args={[
-            Math.max(0.2, objectGroundFootprint(object).width / Math.max(0.001, Math.abs(object.scale[0] || 1))),
-            1,
-            Math.max(0.2, objectGroundFootprint(object).depth / Math.max(0.001, Math.abs(object.scale[2] || 1))),
-          ]} />
-          <meshBasicMaterial transparent opacity={0} depthWrite={false} />
-        </mesh>
-      ) : null}
-    </group>
-  )
-
-  return (
-    <>
-      {selected ? <MannequinFootRings object={object} /> : null}
-      {object.type === 'mannequin' && roleLabel ? <MannequinRoleLabel position={singleMannequinLabelPosition(object)} label={roleLabel} /> : null}
-      {object.type === 'mannequinCrowd' && roleStartIndex !== undefined
-        ? crowdLabelPositions(object).map((position, index) => (
-          <MannequinRoleLabel
-            key={`${object.id}-role-${index}`}
-            position={position}
-            label={mannequinRoleLabel(roleStartIndex + index)}
-          />
-        ))
-        : null}
-      <group ref={anchorRef} position={anchorPosition} rotation={object.rotation} scale={object.scale} />
-      {group}
-      {selected && !readOnly ? (
-        <TransformControls
-          ref={transformRef}
-          object={anchorRef}
-          mode={transformMode}
-          onMouseDown={handleTransformMouseDown}
-          onMouseUp={handleTransformMouseUp}
-          onObjectChange={handleObjectChange}
-        />
-      ) : null}
-    </>
-  )
-}
-
-function CameraHelperView({
-  cameraData,
-  selected,
-  readOnly,
-  orbitControlsActive,
-  navigationLockedRef,
-  onSelect,
-  onFocus,
-  onTransformStart,
-  onTransformEnd,
-  onTransform,
-}: {
-  cameraData: Scene3DCamera
-  selected: boolean
-  readOnly: boolean
-  orbitControlsActive: boolean
-  navigationLockedRef: React.MutableRefObject<boolean>
-  onSelect: () => void
-  onFocus: () => void
-  onTransformStart: () => void
-  onTransformEnd: () => void
-  onTransform: (patch: Partial<Scene3DCamera>) => void
-}): JSX.Element {
-  const markerRef = React.useRef<THREE.Group>(null)
-  const positionDraggingRef = React.useRef(false)
-  const aimDraggingRef = React.useRef<{
-    pointerId: number
-    startX: number
-    startY: number
-    theta: number
-    phi: number
-    radius: number
-    target: PointerCaptureTarget | null
-  } | null>(null)
-  const controlsEnabledBeforeDragRef = React.useRef<boolean | null>(null)
-  const orbitControlsActiveRef = React.useRef(orbitControlsActive)
-  const dragPlaneRef = React.useRef(new THREE.Plane())
-  const dragHitRef = React.useRef(new THREE.Vector3())
-  const dragOffsetRef = React.useRef(new THREE.Vector3())
-  const { controls } = useThree()
-  const target = cameraData.target || CAMERA_DEFAULT_TARGET
-  const cameraPosition = React.useMemo(() => vectorFromArray(cameraData.position), [cameraData.position])
-  const cameraRotation = React.useMemo(
-    () => cameraLookAtRotation(cameraData.position, target),
-    [cameraData.position, target],
-  )
-
-  React.useEffect(() => () => {
-    navigationLockedRef.current = false
-    if (controls && 'enabled' in controls && controlsEnabledBeforeDragRef.current !== null) {
-      ;(controls as { enabled: boolean }).enabled = orbitControlsActiveRef.current
-        ? controlsEnabledBeforeDragRef.current
-        : false
-    }
-  }, [controls, navigationLockedRef])
-
-  React.useLayoutEffect(() => {
-    orbitControlsActiveRef.current = orbitControlsActive
-    if (!orbitControlsActive && controls && 'enabled' in controls && controlsEnabledBeforeDragRef.current === null) {
-      ;(controls as { enabled: boolean }).enabled = false
-    }
-  }, [controls, orbitControlsActive])
-
-  const setSceneControlsDragging = React.useCallback((dragging: boolean) => {
-    navigationLockedRef.current = dragging
-    if (!controls || !('enabled' in controls)) return
-    const orbitControls = controls as { enabled: boolean }
-    if (dragging) {
-      if (controlsEnabledBeforeDragRef.current === null) {
-        controlsEnabledBeforeDragRef.current = orbitControls.enabled
-      }
-      orbitControls.enabled = false
-      return
-    }
-    if (controlsEnabledBeforeDragRef.current !== null) {
-      orbitControls.enabled = orbitControlsActiveRef.current ? controlsEnabledBeforeDragRef.current : false
-      controlsEnabledBeforeDragRef.current = null
-    }
-  }, [controls, navigationLockedRef])
-
-  const stopScenePointerEvent = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    event.nativeEvent.preventDefault()
-    event.nativeEvent.stopPropagation()
-    event.nativeEvent.stopImmediatePropagation()
-    event.stopPropagation()
-  }, [])
-
-  const updatePositionFromEvent = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    const hit = event.ray.intersectPlane(dragPlaneRef.current, dragHitRef.current)
-    if (!hit) return
-    const nextPosition = vectorToArray(hit.clone().add(dragOffsetRef.current))
-    onTransform({
-      position: nextPosition,
-      rotation: cameraLookAtRotation(nextPosition, target),
-    })
-  }, [onTransform, target])
-
-  const handlePositionPointerDown = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    stopScenePointerEvent(event)
-    onSelect()
-    orbitControlsActiveRef.current = false
-    if (readOnly) return
-    onTransformStart()
-    setSceneControlsDragging(true)
-    const planeNormal = new THREE.Vector3()
-    event.camera.getWorldDirection(planeNormal)
-    planeNormal.normalize()
-    dragPlaneRef.current.setFromNormalAndCoplanarPoint(planeNormal, cameraPosition)
-    const hit = event.ray.intersectPlane(dragPlaneRef.current, dragHitRef.current)
-    dragOffsetRef.current.copy(hit ? cameraPosition.clone().sub(hit) : new THREE.Vector3())
-    positionDraggingRef.current = true
-    pointerCaptureTarget(event.target)?.setPointerCapture?.(event.pointerId)
-  }, [cameraPosition, onSelect, onTransformStart, readOnly, setSceneControlsDragging, stopScenePointerEvent])
-
-  const handlePositionPointerMove = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    if (!positionDraggingRef.current || readOnly) return
-    stopScenePointerEvent(event)
-    updatePositionFromEvent(event)
-  }, [readOnly, stopScenePointerEvent, updatePositionFromEvent])
-
-  const stopCameraDrag = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    if (!positionDraggingRef.current) return
-    stopScenePointerEvent(event)
-    positionDraggingRef.current = false
-    setSceneControlsDragging(false)
-    onTransformEnd()
-    pointerCaptureTarget(event.target)?.releasePointerCapture?.(event.pointerId)
-  }, [onTransformEnd, setSceneControlsDragging, stopScenePointerEvent])
-
-  const updateAimFromDrag = React.useCallback((drag: NonNullable<typeof aimDraggingRef.current>, dx: number, dy: number, fine = false) => {
-    const sensitivity = fine ? 0.003 : 0.008
-    const phi = THREE.MathUtils.clamp(drag.phi - dy * sensitivity, 0.08, Math.PI - 0.08)
-    const theta = drag.theta + dx * sensitivity
-    const position = vectorFromArray(cameraData.position)
-    const direction = new THREE.Vector3().setFromSpherical(new THREE.Spherical(drag.radius, phi, theta))
-    const nextTarget = vectorToArray(position.clone().add(direction))
-    onTransform({
-      target: nextTarget,
-      rotation: cameraLookAtRotation(cameraData.position, nextTarget),
-    })
-  }, [cameraData.position, onTransform])
-
-  const handleAimPointerDown = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    stopScenePointerEvent(event)
-    onSelect()
-    orbitControlsActiveRef.current = false
-    if (readOnly) return
-    onTransformStart()
-    const spherical = cameraAimSpherical(cameraData)
-    aimDraggingRef.current = {
-      pointerId: event.pointerId,
-      startX: event.nativeEvent.clientX,
-      startY: event.nativeEvent.clientY,
-      theta: spherical.theta,
-      phi: spherical.phi,
-      radius: Math.max(0.75, spherical.radius),
-      target: pointerCaptureTarget(event.target),
-    }
-    setSceneControlsDragging(true)
-    pointerCaptureTarget(event.target)?.setPointerCapture?.(event.pointerId)
-  }, [cameraData, onSelect, onTransformStart, readOnly, setSceneControlsDragging, stopScenePointerEvent])
-
-  const handleAimPointerMove = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    const drag = aimDraggingRef.current
-    if (!drag || drag.pointerId !== event.pointerId || readOnly) return
-    stopScenePointerEvent(event)
-    updateAimFromDrag(
-      drag,
-      event.nativeEvent.clientX - drag.startX,
-      event.nativeEvent.clientY - drag.startY,
-      event.nativeEvent.shiftKey,
-    )
-  }, [readOnly, stopScenePointerEvent, updateAimFromDrag])
-
-  const stopAimDrag = React.useCallback((event: ThreeEvent<PointerEvent>) => {
-    const drag = aimDraggingRef.current
-    if (!drag || drag.pointerId !== event.pointerId) return
-    stopScenePointerEvent(event)
-    aimDraggingRef.current = null
-    setSceneControlsDragging(false)
-    onTransformEnd()
-    pointerCaptureTarget(event.target)?.releasePointerCapture?.(event.pointerId)
-  }, [onTransformEnd, setSceneControlsDragging, stopScenePointerEvent])
-
-  React.useEffect(() => {
-    const stopNativePointerEvent = (event: PointerEvent) => {
-      event.preventDefault()
-      event.stopPropagation()
-      event.stopImmediatePropagation()
-    }
-
-    const handleWindowPointerMove = (event: PointerEvent) => {
-      const drag = aimDraggingRef.current
-      if (!drag || drag.pointerId !== event.pointerId || readOnly) return
-      stopNativePointerEvent(event)
-      updateAimFromDrag(
-        drag,
-        event.clientX - drag.startX,
-        event.clientY - drag.startY,
-        event.shiftKey,
-      )
-    }
-
-    const stopWindowAimDrag = (event: PointerEvent) => {
-      const drag = aimDraggingRef.current
-      if (!drag || drag.pointerId !== event.pointerId) return
-      stopNativePointerEvent(event)
-      aimDraggingRef.current = null
-      setSceneControlsDragging(false)
-      drag.target?.releasePointerCapture?.(drag.pointerId)
-    }
-
-    window.addEventListener('pointermove', handleWindowPointerMove, { capture: true })
-    window.addEventListener('pointerup', stopWindowAimDrag, { capture: true })
-    window.addEventListener('pointercancel', stopWindowAimDrag, { capture: true })
-    return () => {
-      window.removeEventListener('pointermove', handleWindowPointerMove, { capture: true })
-      window.removeEventListener('pointerup', stopWindowAimDrag, { capture: true })
-      window.removeEventListener('pointercancel', stopWindowAimDrag, { capture: true })
-    }
-  }, [readOnly, setSceneControlsDragging, updateAimFromDrag])
-
-  const marker = (
-    <group
-      ref={markerRef}
-      userData={{ [CAMERA_HELPER_FLAG]: true }}
-      visible={cameraData.visible}
-      position={cameraData.position}
-      rotation={cameraRotation}
-      onPointerDown={handlePositionPointerDown}
-      onPointerMove={handlePositionPointerMove}
-      onPointerUp={stopCameraDrag}
-      onPointerCancel={stopCameraDrag}
-      onDoubleClick={(event) => {
-        event.stopPropagation()
-        onSelect()
-        onFocus()
-      }}
-    >
-      <CameraFrustumLines cameraData={cameraData} selected={selected} />
-      {selected && !readOnly ? (
-        <group
-          position={[0, 0, -CAMERA_AIM_HANDLE_DISTANCE]}
-          onPointerDown={handleAimPointerDown}
-          onPointerMove={handleAimPointerMove}
-          onPointerUp={stopAimDrag}
-          onPointerCancel={stopAimDrag}
-        >
-          <lineSegments frustumCulled={false} raycast={() => null}>
-            <bufferGeometry>
-              <bufferAttribute
-                attach="attributes-position"
-                args={[new Float32Array([
-                  -0.14, 0, 0,
-                  0.14, 0, 0,
-                  0, -0.14, 0,
-                  0, 0.14, 0,
-                ]), 3]}
-              />
-            </bufferGeometry>
-            <lineBasicMaterial color="#facc15" opacity={0.8} transparent toneMapped={false} />
-          </lineSegments>
-          <mesh>
-            <sphereGeometry args={[0.075, 18, 12]} />
-            <meshBasicMaterial color="#facc15" toneMapped={false} />
-          </mesh>
-        </group>
-      ) : null}
-      <mesh>
-        <sphereGeometry args={[0.38, 16, 12]} />
-        <meshBasicMaterial transparent opacity={0} depthWrite={false} />
-      </mesh>
-      <mesh>
-        <boxGeometry args={[0.14, 0.09, 0.08]} />
-        <meshBasicMaterial
-          color={selected ? '#facc15' : CAMERA_MARKER_COLOR}
-          depthWrite={false}
-          opacity={selected ? 0.92 : 0.58}
-          transparent
-          toneMapped={false}
-        />
-      </mesh>
-      <mesh position={[0, 0, -0.12]} rotation={[-Math.PI / 2, 0, 0]}>
-        <coneGeometry args={[0.045, 0.09, 18]} />
-        <meshBasicMaterial
-          color={selected ? '#facc15' : CAMERA_MARKER_ACCENT_COLOR}
-          depthWrite={false}
-          opacity={selected ? 0.92 : 0.58}
-          transparent
-          toneMapped={false}
-        />
-      </mesh>
-    </group>
-  )
-
-  return (
-    <>
-      {marker}
-      {selected ? <CameraTargetFeedback cameraData={cameraData} /> : null}
-    </>
-  )
-}
-
-function SceneContent({
-  state,
-  selection,
-  readOnly,
-  transformMode,
-  flySpeed,
-  focusId,
-  viewLocked,
-  cameraViewEditCamera,
-  onSelect,
-  onFocus,
-  onObjectPatch,
-  onCameraPatch,
-  onEditorCameraDraft,
-  onEditorCameraCommit,
-  onEditorCameraTargetChange,
-  onWheelNavigation,
-  onTransformInteractionStart,
-  onTransformInteractionEnd,
-  onFocusConsumed,
-  onKeyboardNavigationStart,
-  onKeyboardNavigationStop,
-  setCaptureApi,
-}: {
-  state: Scene3DState
-  selection: Scene3DSelection
-  readOnly: boolean
-  transformMode: Scene3DTransformMode
-  flySpeed: number
-  focusId: string
-  viewLocked: boolean
-  cameraViewEditCamera?: Scene3DCamera
-  onSelect: (selection: Scene3DSelection) => void
-  onFocus: (id: string) => void
-  onObjectPatch: (id: string, patch: Partial<Scene3DObject>) => void
-  onCameraPatch: (id: string, patch: Partial<Scene3DCamera>) => void
-  onEditorCameraDraft: (cameraState: Scene3DState['editorCamera']) => void
-  onEditorCameraCommit: (cameraState: Scene3DState['editorCamera']) => void
-  onEditorCameraTargetChange: (target: Scene3DVector3) => void
-  onWheelNavigation: (cameraState: Scene3DState['editorCamera']) => void
-  onTransformInteractionStart: () => void
-  onTransformInteractionEnd: () => void
-  onFocusConsumed: () => void
-  onKeyboardNavigationStart: () => void
-  onKeyboardNavigationStop: () => void
-  setCaptureApi: (api: CaptureApi | null) => void
-}): JSX.Element {
-  const freeLook = !viewLocked
-  const controlMode: Scene3DControlMode = freeLook ? 'fly' : 'edit'
-  const cameraViewEditing = Boolean(cameraViewEditCamera)
-  const navigationLockedRef = React.useRef(false)
-  const mannequinRoleData = React.useMemo(() => {
-    const labels = new Map<string, string>()
-    const starts = new Map<string, number>()
-    let index = 0
-    state.objects.forEach((object) => {
-      if (object.type === 'mannequin') {
-        labels.set(object.id, mannequinRoleLabel(index))
-        starts.set(object.id, index)
-        index += 1
-        return
-      }
-      if (object.type === 'mannequinCrowd') {
-        starts.set(object.id, index)
-        index += crowdCount(object)
-      }
-    })
-    return { labels, starts }
-  }, [state.objects])
-  const gridCellColor = state.environment.darkMode ? DARK_GRID_CELL_COLOR : GRID_CELL_COLOR
-  const gridSectionColor = state.environment.darkMode ? DARK_GRID_SECTION_COLOR : GRID_SECTION_COLOR
-
-  return (
-    <>
-      <color attach="background" args={[state.environment.backgroundColor]} />
-      <ambientLight intensity={0.65} />
-      {state.environment.showSky ? <Sky sunPosition={[2, 1, 4]} /> : null}
-      {state.environment.preset ? (
-        <React.Suspense fallback={null}>
-          <Environment preset="city" />
-        </React.Suspense>
-      ) : null}
-      {state.environment.showGrid && !cameraViewEditing ? (
-        <group userData={{ [SCENE3D_GRID_FLAG]: true }}>
-          <Grid
-            infiniteGrid
-            cellSize={0.5}
-            sectionSize={5}
-            fadeDistance={42}
-            fadeStrength={1.25}
-            cellColor={gridCellColor}
-            sectionColor={gridSectionColor}
-          />
-        </group>
-      ) : null}
-      {state.environment.showAxes && !cameraViewEditing ? <axesHelper args={[2]} /> : null}
-      {state.objects.map((object) => (
-        <SceneObjectView
-          key={object.id}
-          object={object}
-          selected={selection?.type === 'object' && selection.id === object.id}
-          readOnly={readOnly}
-          transformMode={transformMode}
-          orbitControlsActive={!freeLook}
-          navigationLockedRef={navigationLockedRef}
-          roleLabel={object.type === 'mannequin' ? mannequinRoleData.labels.get(object.id) : undefined}
-          roleStartIndex={mannequinRoleData.starts.get(object.id)}
-          onSelect={() => onSelect({ type: 'object', id: object.id })}
-          onFocus={() => onFocus(object.id)}
-          onTransformStart={onTransformInteractionStart}
-          onTransformEnd={onTransformInteractionEnd}
-          onTransform={(patch) => onObjectPatch(object.id, patch)}
-        />
-      ))}
-      {!cameraViewEditing ? state.cameras.map((camera) => (
-        <CameraHelperView
-          key={camera.id}
-          cameraData={camera}
-          selected={selection?.type === 'camera' && selection.id === camera.id}
-          readOnly={readOnly}
-          orbitControlsActive={!freeLook}
-          navigationLockedRef={navigationLockedRef}
-          onSelect={() => onSelect({ type: 'camera', id: camera.id })}
-          onFocus={() => onFocus(camera.id)}
-          onTransformStart={onTransformInteractionStart}
-          onTransformEnd={onTransformInteractionEnd}
-          onTransform={(patch) => onCameraPatch(camera.id, patch)}
-        />
-      )) : null}
-      <InitialCameraPose editorCamera={state.editorCamera} />
-      <CameraViewEditController
-        cameraData={cameraViewEditCamera}
-        onCameraPatch={onCameraPatch}
-        onEditorCameraDraft={onEditorCameraDraft}
-      />
-      <FocusController
-        focusId={focusId}
-        objects={state.objects}
-        cameras={state.cameras}
-        onTargetChange={onEditorCameraTargetChange}
-        onFocusConsumed={onFocusConsumed}
-      />
-      <Scene3DControls
-        freeLook={freeLook}
-        selectionActive={selection !== null}
-        speed={flySpeed}
-        target={state.editorCamera.target}
-        navigationLockedRef={navigationLockedRef}
-        onClearSelection={() => onSelect(null)}
-        onWheelNavigation={onWheelNavigation}
-        onKeyboardNavigationStart={onKeyboardNavigationStart}
-        onKeyboardNavigationStop={onKeyboardNavigationStop}
-      />
-      <CameraStateRecorder
-        mode={controlMode}
-        target={state.editorCamera.target}
-        onDraftChange={onEditorCameraDraft}
-        onCommit={onEditorCameraCommit}
-      />
-      <CaptureBinder cameras={state.cameras} setApi={setCaptureApi} />
-    </>
-  )
-}
-
-function cameraPreviewViewportStyle(aspectRatio: Scene3DAspectRatio): React.CSSProperties {
-  const ratio = SCENE3D_ASPECT_RATIOS[aspectRatio]
-  const maxWidth = 224
-  const maxHeight = 240
-  let width = maxWidth
-  let height = width / ratio
-  if (height > maxHeight) {
-    height = maxHeight
-    width = height * ratio
-  }
-  return {
-    width: `${Math.round(width)}px`,
-    height: `${Math.round(height)}px`,
-  }
-}
-
-function CameraPreviewPose({ cameraData }: { cameraData: Scene3DCamera }): null {
-  const { camera } = useThree()
-
-  React.useLayoutEffect(() => {
-    applySceneCameraPose(camera, cameraData)
-  }, [camera, cameraData])
-
-  return null
-}
-
-function PreviewObjectView({
-  object,
-  roleStartIndex = 0,
-}: {
-  object: Scene3DObject
-  roleStartIndex?: number
-}): JSX.Element {
-  return (
-    <group
-      visible={object.visible}
-      position={object.position}
-      rotation={object.rotation}
-      scale={object.scale}
-    >
-      {object.type === 'mannequin' ? (
-        <MannequinAssetBoundary fallback={<ProceduralMannequin color={object.color || '#808080'} />}>
-          <React.Suspense fallback={<ProceduralMannequin color={object.color || '#808080'} />}>
-            <Mannequin color={object.color || '#808080'} pose={object.pose} />
-          </React.Suspense>
-        </MannequinAssetBoundary>
-      ) : object.type === 'mannequinCrowd' ? (
-        <ProceduralMannequinCrowd object={object} roleStartIndex={roleStartIndex} />
-      ) : object.type === 'light' ? (
-        <LightObject object={object} />
-      ) : (
-        <mesh>
-          <Scene3DMeshGeometry geometry={object.geometry} />
-          <meshStandardMaterial
-            color={object.color || '#808080'}
-            roughness={0.55}
-            metalness={0.04}
-            side={object.geometry === 'plane' ? THREE.DoubleSide : THREE.FrontSide}
-          />
-        </mesh>
-      )}
-    </group>
-  )
-}
-
-function CameraPreviewScene({
-  state,
-  cameraData,
-}: {
-  state: Scene3DState
-  cameraData: Scene3DCamera
-}): JSX.Element {
-  let roleIndex = 0
-  return (
-    <>
-      <color attach="background" args={[state.environment.backgroundColor]} />
-      <ambientLight intensity={0.65} />
-      {state.environment.showSky ? <Sky sunPosition={[2, 1, 4]} /> : null}
-      {state.environment.preset ? (
-        <React.Suspense fallback={null}>
-          <Environment preset="city" />
-        </React.Suspense>
-      ) : null}
-      {state.environment.showAxes ? <axesHelper args={[2]} /> : null}
-      {state.objects.map((object) => {
-        const roleStartIndex = roleIndex
-        if (object.type === 'mannequin') roleIndex += 1
-        if (object.type === 'mannequinCrowd') roleIndex += crowdCount(object)
-        return <PreviewObjectView key={object.id} object={object} roleStartIndex={roleStartIndex} />
-      })}
-      <CameraPreviewPose cameraData={cameraData} />
-    </>
-  )
-}
-
-function CameraPreview({
-  camera,
-  state,
-  readOnly,
-  cameraViewEditing,
-  rightPanelCollapsed,
-  onAspectChange,
-  onLensDepthChange,
-  onToggleViewEdit,
-  onLevelCamera,
-  onScreenshot,
-}: {
-  camera: Scene3DCamera
-  state: Scene3DState
-  readOnly: boolean
-  cameraViewEditing: boolean
-  rightPanelCollapsed: boolean
-  onAspectChange: (aspectRatio: Scene3DAspectRatio) => void
-  onLensDepthChange: (lensDepth: number) => void
-  onToggleViewEdit: () => void
-  onLevelCamera: () => void
-  onScreenshot: () => void
-}): JSX.Element {
-  const previewStyle = React.useMemo(() => cameraPreviewViewportStyle(camera.aspectRatio), [camera.aspectRatio])
-  const lensDepth = camera.lensDepth ?? 0
-
-  return (
-    <div
-      className={cn(
-        'absolute right-4 z-[3] w-[260px] rounded-nomi border border-[var(--nomi-line-soft)] bg-[var(--nomi-paper)] p-2 text-[var(--nomi-ink)] shadow-[var(--nomi-shadow-md)]',
-        rightPanelCollapsed ? 'top-16' : 'top-4',
-      )}
-    >
-      <div className="mb-2 flex items-center justify-between gap-2">
-        <div className="min-w-0 truncate text-caption font-medium">{camera.name} · {camera.aspectRatio}</div>
-        <div className="flex shrink-0 items-center gap-1">
-          <button
-            className={cn(
-              'inline-flex h-7 items-center gap-1 rounded-nomi-sm px-2 text-micro hover:bg-[var(--nomi-ink-10)] hover:text-[var(--nomi-ink)] disabled:opacity-40',
-              cameraViewEditing ? 'bg-[var(--nomi-ink)] text-[var(--nomi-paper)]' : 'bg-[var(--nomi-ink-05)] text-[var(--nomi-ink-60)]',
-            )}
-            disabled={readOnly}
-            type="button"
-            title={cameraViewEditing ? '正在取景调整，按 Esc 或点击顶部退出' : '从相机视角调整'}
-            onClick={onToggleViewEdit}
-          >
-            <IconEye size={14} />
-            <span>取景</span>
-          </button>
-          <button
-            className="grid size-7 place-items-center rounded-nomi-sm bg-[var(--nomi-ink-05)] text-[var(--nomi-ink-60)] hover:bg-[var(--nomi-ink-10)] hover:text-[var(--nomi-ink)] disabled:opacity-40"
-            disabled={readOnly}
-            type="button"
-            title="水平摆正"
-            onClick={onLevelCamera}
-          >
-            <IconRotate size={14} />
-          </button>
-          <button className="grid size-7 place-items-center rounded-nomi-sm bg-[var(--nomi-ink-05)] text-[var(--nomi-ink-60)] hover:bg-[var(--nomi-ink-10)] hover:text-[var(--nomi-ink)]" type="button" title="相机截图" onClick={onScreenshot}>
-            <IconCamera size={15} />
-          </button>
-        </div>
-      </div>
-      <div className="flex min-h-[126px] items-center justify-center rounded-nomi-sm border border-[var(--nomi-line-soft)] bg-[var(--nomi-ink-05)] p-1">
-        <div className="overflow-hidden rounded-nomi-sm bg-[var(--nomi-ink)]" style={previewStyle}>
-          <Canvas
-            camera={{
-              fov: camera.fov,
-              near: camera.near,
-              far: camera.far,
-              position: camera.position,
-              rotation: camera.rotation,
-            }}
-            dpr={[1, 1.5]}
-            frameloop="demand"
-            gl={{ antialias: true, preserveDrawingBuffer: false }}
-          >
-            <CameraPreviewScene state={state} cameraData={camera} />
-          </Canvas>
-        </div>
-      </div>
-      <div className="mt-2 grid grid-cols-5 gap-1">
-        {SCENE3D_ASPECT_OPTIONS.map((option) => (
-          <button
-            key={option}
-            className={cn(
-              'h-6 rounded-nomi-sm border border-[var(--nomi-line-soft)] text-micro text-[var(--nomi-ink-60)] hover:bg-[var(--nomi-ink-05)] hover:text-[var(--nomi-ink)]',
-              option === camera.aspectRatio && 'bg-[var(--nomi-ink)] text-[var(--nomi-paper)]',
-            )}
-            disabled={readOnly}
-            type="button"
-            onClick={() => onAspectChange(option)}
-          >
-            {option}
-          </button>
-        ))}
-      </div>
-      <div className="mt-3 rounded-nomi-sm border border-[var(--nomi-line-soft)] bg-[var(--nomi-ink-05)] px-2 py-2">
-        <div className="mb-1 flex items-center justify-between gap-2 text-micro text-[var(--nomi-ink-60)]">
-          <span>镜头深度</span>
-          <span className="font-medium text-[var(--nomi-ink)]">{Math.round(lensDepth)}%</span>
-        </div>
-        <input
-          className="block h-1.5 w-full accent-[var(--nomi-ink)]"
-          disabled={readOnly}
-          max={100}
-          min={-100}
-          step={1}
-          type="range"
-          value={lensDepth}
-          onChange={(event) => onLensDepthChange(Number(event.currentTarget.value))}
-        />
-        <div className="mt-1 grid grid-cols-3 text-micro text-[var(--nomi-ink-45)]">
-          <span>-100%</span>
-          <span className="text-center">0</span>
-          <span className="text-right">100%</span>
-        </div>
-      </div>
-    </div>
-  )
-}
 
 export default function Scene3DFullscreen({
   initialState,
